Rename sample to template
