--- src/main/java/org/wildstang/sample/robot/WsOutputs.java
@@ -1,84 +1,0 @@
-package org.wildstang.sample.robot;
-
-import org.wildstang.framework.core.Core;
-import org.wildstang.framework.core.Outputs;
-import org.wildstang.framework.hardware.OutputConfig;
-import org.wildstang.framework.io.outputs.Output;
-import org.wildstang.hardware.roborio.outputs.config.WsSparkMaxConfig;
-
-/**
- * Output mappings are stored here.
- * Below each Motor, Solenoid, etc is enumerated with their appropriated IDs.
- * The enumeration includes a name, output type, and output config object.
- */
-public enum WsOutputs implements Outputs {
-
-    // ---------------------------------
-    // Motors
-    // ---------------------------------
-    DRIVE1("Module 1 Drive Motor", new WsSparkMaxConfig(CANConstants.DRIVE1, true)),
-    ANGLE1("Module 1 Angle Motor", new WsSparkMaxConfig(CANConstants.ANGLE1, true)),
-    DRIVE2("Module 2 Drive Motor", new WsSparkMaxConfig(CANConstants.DRIVE2, true)),
-    ANGLE2("Module 2 Angle Motor", new WsSparkMaxConfig(CANConstants.ANGLE2, true)),
-    DRIVE3("Module 3 Drive Motor", new WsSparkMaxConfig(CANConstants.DRIVE3, true)),
-    ANGLE3("Module 3 Angle Motor", new WsSparkMaxConfig(CANConstants.ANGLE3, true)),
-    DRIVE4("Module 4 Drive Motor", new WsSparkMaxConfig(CANConstants.DRIVE4, true)),
-    ANGLE4("Module 4 Angle Motor", new WsSparkMaxConfig(CANConstants.ANGLE4, true)),
-
-    // ---------------------------------
-    // Solenoids
-    // ---------------------------------
-
-    // ---------------------------------
-    // Other
-    // ---------------------------------
-
-    ; // end of enum
-
-    /**
-     * Do not modify below code, provides template for enumerations.
-     * We would like to have a super class for this structure, however,
-     * Java does not support enums extending classes.
-     */
-
-    private String m_name;
-    private OutputConfig m_config;
-
-    /**
-     * Initialize a new Output.
-     * 
-     * @param p_name   Name, must match that in class to prevent errors.
-     * @param p_config Corresponding configuration for OutputType.
-     */
-    WsOutputs(String p_name, OutputConfig p_config) {
-        m_name = p_name;
-        m_config = p_config;
-    }
-
-    /**
-     * Returns the name mapped to the Output.
-     * 
-     * @return Name mapped to the Output.
-     */
-    public String getName() {
-        return m_name;
-    }
-
-    /**
-     * Returns the config of Output for the enumeration.
-     * 
-     * @return OutputConfig of enumeration.
-     */
-    public OutputConfig getConfig() {
-        return m_config;
-    }
-
-    /**
-     * Returns the actual Output object from the OutputManager
-     * 
-     * @return The corresponding output.
-     */
-    public Output get() {
-        return Core.getOutputManager().getOutput(this);
-    }
-}
