+++ src/main/java/org/wildstang/template/subsystems/drive/DriveSignal.java
@@ -1,0 +1,29 @@
+package org.wildstang.template.subsystems.drive;
+
+/**
+ * A drivetrain command consisting of the left, right motor settings and whether
+ * the brake mode is enabled.
+ */
+public class DriveSignal {
+    public double leftMotor;
+    public double rightMotor;
+    public boolean brakeMode;
+
+    public DriveSignal(double left, double right) {
+        this(left, right, false);
+    }
+
+    public DriveSignal(double left, double right, boolean brakeMode) {
+        this.leftMotor = left;
+        this.rightMotor = right;
+        this.brakeMode = brakeMode;
+    }
+
+    public static DriveSignal NEUTRAL = new DriveSignal(0, 0);
+    public static DriveSignal BRAKE = new DriveSignal(0, 0, true);
+
+    @Override
+    public String toString() {
+        return "L: " + leftMotor + ", R: " + rightMotor;
+    }
+}
